+++ app/Exceptions/Handler.php
@@ -71,7 +71,7 @@ class Handler extends ExceptionHandler
         if ($request->expectsJson()) {
             return response()->json(['error' => 'Unauthenticated.'], 401);
         } else {
-            return redirect()->guest('login');
+            return view('auth.login');
         }
     }
 }
