--- app/Http/Controllers/Auth/AuthController.php
@@ -1,102 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use Illuminate\Http\Request;
-use Session;
-use Validator;
-use App\Models\User;
-use App\Http\Controllers\Controller;
-use Illuminate\Foundation\Auth\AuthenticatesUsers;
-
-class AuthController extends Controller
-{
-    /*
-    |--------------------------------------------------------------------------
-    | Registration & Login Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller handles the registration of new users, as well as the
-    | authentication of existing users. By default, this controller uses
-    | a simple trait to add these behaviors. Why don't you explore it?
-    |
-    */
-    use AuthenticatesUsers;
-
-    /**
-     * @var string
-     */
-    protected $redirectAfterLogout = 'auth/login';
-
-    /**
-     * @var string
-     */
-    protected $redirectTo = 'admin';
-
-    /**
-     * Create a new authentication controller instance.
-     */
-    public function __construct()
-    {
-        $this->middleware('guest', ['except' => 'logout']);
-    }
-
-    /**
-     * Log the user out of the application.
-     *
-     * @param  Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function logout(Request $request)
-    {
-        $this->guard()->logout();
-
-        $request->session()->flush();
-
-        $request->session()->regenerate();
-
-        return redirect('/auth/login');
-    }
-
-    /**
-     * Get a validator for an incoming registration request.
-     *
-     * @param  array $data
-     * @return \Illuminate\Contracts\Validation\Validator
-     */
-    protected function validator(array $data)
-    {
-        return Validator::make($data, [
-            'name' => 'required|max:255',
-            'email' => 'required|email|max:255|unique:users',
-            'password' => 'required|confirmed|min:6',
-        ]);
-    }
-
-    /**
-     * Create a new user instance after a valid registration.
-     *
-     * @param  array $data
-     * @return User
-     */
-    protected function create(array $data)
-    {
-        return User::create([
-            'name' => $data['name'],
-            'email' => $data['email'],
-            'password' => bcrypt($data['password']),
-        ]);
-    }
-
-    /**
-     * @param \Illuminate\Http\Request $request
-     * @param User $user
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function authenticated(\Illuminate\Http\Request $request, User $user)
-    {
-        Session::set('_login', trans('messages.login', ['first_name' => $user->first_name, 'last_name' => $user->last_name]));
-
-        return redirect()->intended($this->redirectPath());
-    }
-}
